cleanup BadukWithAbstractBoard, move files, fix imports
