+++ packages/shared/src/lib/abstractBoard/helper/types/intersection.ts
@@ -1,15 +1,13 @@
-import { Color, StoneState } from "./StoneState";
 import type { Vector2D } from "./Vector2D";
 
+// class is used in the board creation functions
 export class Intersection {
   Neighbours: Intersection[];
   Position: Vector2D;
-  StoneState: StoneState;
 
   constructor(v: Vector2D) {
     this.Position = v;
     this.Neighbours = [];
-    this.StoneState = new StoneState(Color.EMPTY);
   }
 
   ConnectTo(intersection: Intersection, bothSides: boolean) {
@@ -25,7 +23,6 @@ export class Intersection {
 
   Export(): Intersection {
     const intersection = new Intersection(this.Position.Export());
-    intersection.StoneState = this.StoneState.Export();
     //ToDo: How to export Neighbours?
     return intersection;
   }
