+++ packages/shared/src/variant_map.ts
@@ -1,5 +1,5 @@
 import { AbstractGame } from "./abstract_game";
-import { badukWithAbstractBoardVariant } from "./variants/badukWithAbstractBoard";
+import { badukWithAbstractBoardVariant } from "./variants/deprecated/badukWithAbstractBoard";
 import { phantomVariant } from "./variants/phantom";
 import { parallelVariant } from "./variants/parallel";
 import { captureVariant } from "./variants/capture";
