--- packages/shared/src/variants/badukWithAbstractBoard/abstractBoard/Interfaces/IBadukBoard.ts
@@ -1,5 +1,0 @@
-import type { Intersection } from "../intersection";
-
-export interface IBadukBoard {
-  Intersections: Intersection[];
-}
