--- packages/shared/src/variants/badukWithAbstractBoard/abstractBoard/StoneState.ts
@@ -1,17 +1,0 @@
-export enum Color {
-  EMPTY = 0,
-  BLACK = 1,
-  WHITE = 2,
-}
-
-export class StoneState {
-  Color: Color;
-
-  constructor(_color: Color) {
-    this.Color = _color;
-  }
-
-  Export(): StoneState {
-    return new StoneState(this.Color);
-  }
-}
