--- packages/shared/src/variants/badukWithAbstractBoard/index.ts
@@ -1,1 +1,0 @@
-export * from "./badukWithAbstractBoard";
