+++ packages/vue-client/src/components/GameCreation/GameCreationForm.vue
@@ -4,12 +4,12 @@ import {
   getVariantList,
   getDefaultConfig,
   type ITimeControlConfig,
+  BoardPattern,
 } from "@ogfcommunity/variants-shared";
 import * as requests from "@/requests";
 import router from "@/router";
 import { config_form_map } from "@/config_form_map";
 import TimeControlConfigForm from "./TimeControlConfigForm.vue";
-import { BoardPattern } from "@ogfcommunity/variants-shared/src/variants/badukWithAbstractBoard";
 
 interface IConfigWithOptionalTimeControl {
   time_control?: ITimeControlConfig;
