+++ packages/vue-client/src/config_form_map.ts
@@ -1,6 +1,5 @@
 import type { Component } from "vue";
 import BadukConfigForm from "@/components/GameCreation/BadukConfigForm.vue";
-import BadukWithAbstractBoardConfigForm from "@/components/GameCreation/BadukWithAbstractBoardConfigForm.vue";
 import ParalleGoConfigForm from "@/components/GameCreation/ParallelGoConfigForm.vue";
 import DriftGoConfigFormVue from "./components/GameCreation/DriftGoConfigForm.vue";
 import GridBadukConfigForm from "./components/GameCreation/GridBadukConfigForm.vue";
@@ -11,7 +10,6 @@ export const config_form_map: {
 } = {
   baduk: BadukConfigForm,
   phantom: BadukConfigForm,
-  badukWithAbstractBoard: BadukWithAbstractBoardConfigForm,
   parallel: ParalleGoConfigForm,
   capture: BadukConfigForm,
   tetris: BadukConfigForm,
